Use modern i18n mechanisms for localization
The LanguageGetMagic hook was deprecated in 1.16, so move to modern
versions of localization. The parser functions can easily be moved to an
i18n.magic.php file.

There's currently no standard mechanism to localize tag hooks or
attributes, but it can at least be moved out of the $wgLstLocal global
into a private static class attribute. Moving it to i18n messages was
considered, but skipped for now due to the danger of someone changing
localizations, breaking existing content.

Bug: T198173
Co-Authored-By: Krinkle <krinklemail@gmail.com>
Change-Id: Iebaac95d0a3e8b81f212d01265793c1b65a32024